fix: guard symlinks and grep-q in build_latest_perl.sh and build_imagick.sh

+++ scripts/build_latest_perl.sh
@@ -40,7 +40,8 @@ fi
 
 # Build some symlinks to the perl for use by other? setup scripts
 mkdir -p $CLIENT_HOMEDIR/bin
-ln -s /opt/perl5/$NICE_PERL_NAME/bin/perl  $CLIENT_HOMEDIR/bin/perl
-ln -s /opt/perl5/$NICE_PERL_NAME/bin/cpanm $CLIENT_HOMEDIR/bin/cpanm
-ln -s /opt/perl5/$NICE_PERL_NAME/bin/perl  /root/bin/perl
-ln -s /opt/perl5/$NICE_PERL_NAME/bin/cpanm /root/bin/cpanm
+mkdir -p /root/bin
+[ -L $CLIENT_HOMEDIR/bin/perl  ] || ln -s /opt/perl5/$NICE_PERL_NAME/bin/perl  $CLIENT_HOMEDIR/bin/perl
+[ -L $CLIENT_HOMEDIR/bin/cpanm ] || ln -s /opt/perl5/$NICE_PERL_NAME/bin/cpanm $CLIENT_HOMEDIR/bin/cpanm
+[ -L /root/bin/perl  ] || ln -s /opt/perl5/$NICE_PERL_NAME/bin/perl  /root/bin/perl
+[ -L /root/bin/cpanm ] || ln -s /opt/perl5/$NICE_PERL_NAME/bin/cpanm /root/bin/cpanm
